requirements.txt
tweepy
transformers
torch
faiss-cpu
sentence-transformers
